+++ kanu/utils.py
@@ -1,4 +1,109 @@
 import tkinter as tk
+from tkinter import font
+
+class Settings:
+    def __init__(self, agent):
+        self.default_font = font.nametofont("TkDefaultFont").actual()
+        self.default_user_background_color = "gray85"
+        self.default_user_foreground_color = "black"
+        self.default_user_font_family = self.default_font["family"]
+        self.default_user_font_size = self.default_font["size"]
+        self.default_bot_background_color = "white"
+        self.default_bot_foreground_color = "black"
+        self.default_bot_font_family = self.default_font["family"]
+        self.default_bot_font_size = self.default_font["size"]
+        self.agent = agent
+        self.user_background_color = tk.StringVar(self.agent.kanu.container, value=self.default_user_background_color)
+        self.user_foreground_color = tk.StringVar(self.agent.kanu.container, value=self.default_user_foreground_color)
+        self.user_font_family = tk.StringVar(self.agent.kanu.container, value=self.default_user_font_family)
+        self.user_font_size = tk.IntVar(self.agent.kanu.container, value=self.default_user_font_size)
+        self.bot_background_color = tk.StringVar(self.agent.kanu.container, value=self.default_bot_background_color)
+        self.bot_foreground_color = tk.StringVar(self.agent.kanu.container, value=self.default_bot_foreground_color)
+        self.bot_font_family = tk.StringVar(self.agent.kanu.container, value=self.default_bot_font_family)
+        self.bot_font_size = tk.IntVar(self.agent.kanu.container, value=self.default_bot_font_size)
+
+    def get_user_kwargs(self):
+        return dict(
+            background=self.user_background_color.get(),
+            foreground=self.user_foreground_color.get(),
+            font=(self.user_font_family.get(), self.user_font_size.get())
+        )
+            
+    def get_bot_kwargs(self):
+        return dict(
+            background=self.bot_background_color.get(),
+            foreground=self.bot_foreground_color.get(),
+            font=(self.bot_font_family.get(), self.bot_font_size.get())
+        )
+
+    def page(self):
+        self.agent.previous = self.agent.kanu.container
+        self.agent.kanu.container.pack_forget()
+        self.agent.kanu.container = tk.Frame(self.agent.kanu.root)
+        self.agent.kanu.container.pack()
+        l = tk.Label(self.agent.kanu.container, text=self.agent.__class__.__name__)
+        l.grid(row=0, column=0, columnspan=3)
+        l = tk.Label(self.agent.kanu.container, text="User background color")
+        l.grid(row=1, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.user_background_color)
+        e.grid(row=1, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="User foreground color")
+        l.grid(row=2, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.user_foreground_color)
+        e.grid(row=2, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="User font family")
+        l.grid(row=3, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.user_font_family)
+        e.grid(row=3, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="User font size")
+        l.grid(row=4, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.user_font_size)
+        e.grid(row=4, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="Bot background color")
+        l.grid(row=5, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.bot_background_color)
+        e.grid(row=5, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="Bot foreground color")
+        l.grid(row=6, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.bot_foreground_color)
+        e.grid(row=6, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="Bot font family")
+        l.grid(row=7, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.bot_font_family)
+        e.grid(row=7, column=1, columnspan=2)
+        l = tk.Label(self.agent.kanu.container, text="Bot font size")
+        l.grid(row=8, column=0)
+        e = tk.Entry(self.agent.kanu.container, textvariable=self.bot_font_size)
+        e.grid(row=8, column=1, columnspan=2)
+        b = tk.Button(self.agent.kanu.container, text="Apply", command=lambda: self.apply())
+        b.grid(row=9, column=0)
+        b = tk.Button(self.agent.kanu.container, text="Reset", command=lambda: self.reset())
+        b.grid(row=9, column=1)
+        b = tk.Button(self.agent.kanu.container, text="Go back", command=lambda: self.go_back())
+        b.grid(row=9, column=2)
+
+    def go_back(self):
+        self.agent.kanu.container.pack_forget()
+        self.agent.kanu.container = self.agent.previous
+        self.agent.kanu.container.pack()
+
+    def apply(self):
+        self.agent.session.tag_config("user", **self.get_user_kwargs())
+        self.agent.session.tag_config("bot", **self.get_bot_kwargs())
+        self.agent.kanu.container.pack_forget()
+        self.agent.kanu.container = self.agent.previous
+        self.agent.kanu.container.pack()
+
+    def reset(self):
+        self.user_background_color = tk.StringVar(self.agent.kanu.container, value=self.default_user_background_color)
+        self.user_foreground_color = tk.StringVar(self.agent.kanu.container, value=self.default_user_foreground_color)
+        self.user_font_family = tk.StringVar(self.agent.kanu.container, value=self.default_user_font_family)
+        self.user_font_size = tk.IntVar(self.agent.kanu.container, value=self.default_user_font_size)
+        self.bot_background_color = tk.StringVar(self.agent.kanu.container, value=self.default_bot_background_color)
+        self.bot_foreground_color = tk.StringVar(self.agent.kanu.container, value=self.default_bot_foreground_color)
+        self.bot_font_family = tk.StringVar(self.agent.kanu.container, value=self.default_bot_font_family)
+        self.bot_font_size = tk.IntVar(self.agent.kanu.container, value=self.default_bot_font_size)
+        self.apply()
 
 class Tooltip:
     def __init__(self, widget, text):